feat(capabilities)!: retire platform_management
`platform` supersedes it: the same management surface derived from the
authoritative command catalog instead of 14 hand-written tools that had to be
maintained alongside the API and drifted from it. No built-in harness has
referenced `platform_management` since Platform Chat switched.

The capability is now a `Retired` shell — identity only, no tools, prompt,
mounts, or dependencies — so agents and harnesses that still list it keep
resolving and running while the editors flag it for removal. The docs page
becomes a removal notice with a tool-to-command mapping;
`session_read_response`, which blocked until a turn finished, has no catalog
equivalent and is called out as such.

The embedded product-docs mount moves to `platform`, where the published
capability reference already said it lived. It was orphaned when Platform Chat
switched capabilities, so Platform Chat gains back /workspace/docs and, with it,
`session_file_system` as a dependency.

`PlatformStore`'s legacy org-scoped CRUD stays for now: the delegation
capabilities still call parts of it. Trimming the unused remainder is a
follow-up.
