+++ docs/capabilities/index.md
@@ -94,7 +94,6 @@ Agent self-management and platform control.
 | Capability | ID | Tools |
 |---|---|---|
 | [Platform](/capabilities/platform/) | `platform` | 3 |
-| [Platform Management (legacy)](/capabilities/platform-management/) | `platform_management` | 14 |
 
 ### Optimization
